Revert "adding alternative way of reporting on durable queue messages"
This reverts commit af7d5713c793c020a013fa4a4f04885ef93955bc.
