--- components/foundation/src/main/java/dk/cloudcreate/essentials/components/foundation/messaging/queue/micrometer/DurableQueuesMonitor.java
@@ -1,8 +1,0 @@
-package dk.cloudcreate.essentials.components.foundation.messaging.queue.micrometer;
-
-import dk.cloudcreate.essentials.components.foundation.messaging.queue.QueueName;
-
-public interface DurableQueuesMonitor {
-    void monitor(QueueName queueName);
-    boolean enabled();
-}
