+++ components/spring-boot-starter-mongodb/src/main/java/dk/cloudcreate/essentials/components/boot/autoconfigure/mongodb/EssentialsComponentsProperties.java
@@ -16,15 +16,10 @@
 
 package dk.cloudcreate.essentials.components.boot.autoconfigure.mongodb;
 
-import dk.cloudcreate.essentials.components.distributed.fencedlock.springdata.mongo.MongoFencedLockManager;
-import dk.cloudcreate.essentials.components.distributed.fencedlock.springdata.mongo.MongoFencedLockStorage;
+import dk.cloudcreate.essentials.components.distributed.fencedlock.springdata.mongo.*;
 import dk.cloudcreate.essentials.components.foundation.IOExceptionUtil;
-import dk.cloudcreate.essentials.components.foundation.fencedlock.DBFencedLock;
-import dk.cloudcreate.essentials.components.foundation.fencedlock.DBFencedLockManager;
-import dk.cloudcreate.essentials.components.foundation.fencedlock.FencedLock;
-import dk.cloudcreate.essentials.components.foundation.fencedlock.FencedLockStorage;
-import dk.cloudcreate.essentials.components.foundation.messaging.queue.QueueName;
-import dk.cloudcreate.essentials.components.foundation.messaging.queue.TransactionalMode;
+import dk.cloudcreate.essentials.components.foundation.fencedlock.*;
+import dk.cloudcreate.essentials.components.foundation.messaging.queue.*;
 import dk.cloudcreate.essentials.components.foundation.messaging.queue.operations.ConsumeFromQueue;
 import dk.cloudcreate.essentials.components.foundation.mongo.MongoUtil;
 import dk.cloudcreate.essentials.components.foundation.transaction.UnitOfWork;
@@ -35,8 +30,7 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
 import reactor.core.publisher.Sinks;
 
-import java.time.Duration;
-import java.time.OffsetDateTime;
+import java.time.*;
 
 /**
  * Properties for the MongoDB focused Essentials Components auto-configuration<br>
@@ -89,7 +83,6 @@ public class EssentialsComponentsProperties {
     private final TracingProperties   tracingProperties             = new TracingProperties();
     private       boolean             immutableJacksonModuleEnabled = true;
     private final ReactiveProperties  reactive                      = new ReactiveProperties();
-    private final DurableQueuesMonitorProperties durableQueuesMonitor = new DurableQueuesMonitorProperties();
 
     /**
      * Should the EssentialsImmutableJacksonModule be included in the ObjectMapper configuration - default is true<br>
@@ -129,10 +122,6 @@ public class EssentialsComponentsProperties {
 
     public ReactiveProperties getReactive() {
         return reactive;
-    }
-
-    public DurableQueuesMonitorProperties getDurableQueuesMonitor() {
-        return this.durableQueuesMonitor;
     }
 
     public static class DurableQueues {
@@ -565,68 +554,6 @@ public class EssentialsComponentsProperties {
         public void setQueuedTaskCapFactor(double queuedTaskCapFactor) {
             this.queuedTaskCapFactor = queuedTaskCapFactor;
         }
-    }
-
-    public static class DurableQueuesMonitorProperties {
-        private boolean enabled = false;
-        private Duration interval = Duration.ofMinutes(1);
-        private boolean micrometerMonitorEnabled = false;
-
-        /**
-         * Is monitoring of queued and dead letter messages enabled using implementation of {@link dk.cloudcreate.essentials.components.foundation.messaging.queue.micrometer.DurableQueuesMicrometerMonitor}
-         * This will disable the {@link dk.cloudcreate.essentials.components.foundation.messaging.queue.micrometer.DurableQueuesMicrometerInterceptor} and only report on queues in the given interval
-         * @return Is monitoring of queued and dead letter messages enabled
-         */
-        public boolean isMicrometerMonitorEnabled() {
-            return this.micrometerMonitorEnabled;
-        }
-
-        /**
-         * Is monitoring of queued and dead letter messages enabled using implementation of {@link dk.cloudcreate.essentials.components.foundation.messaging.queue.micrometer.DurableQueuesMicrometerMonitor}
-         * This will disable the {@link dk.cloudcreate.essentials.components.foundation.messaging.queue.micrometer.DurableQueuesMicrometerInterceptor} and only report on queues in the given interval
-         * @param micrometerMonitorEnabled Is monitoring of queued and dead letter messages enabled
-         */
-        public void setMicrometerMonitorEnabled(boolean micrometerMonitorEnabled) {
-            this.micrometerMonitorEnabled = micrometerMonitorEnabled;
-        }
-
-        /**
-         * Is monitoring of durable queues enabled
-         * @return Is monitoring of durable queues enabled
-         */
-        public boolean isEnabled() {
-            return this.enabled;
-        }
-
-        /**
-         * Monitoring of durable queues using implementations of
-         * {@link dk.cloudcreate.essentials.components.foundation.messaging.queue.micrometer.DurableQueuesMonitor}
-         * @param enabled Monitoring of durable queues
-         */
-        public void setEnabled(boolean enabled) {
-            this.enabled = enabled;
-        }
-
-        /**
-         * Monitoring interval
-         * @param interval Monitoring interval
-         */
-        public void setInterval(Duration interval) {
-            this.interval = interval;
-        }
-
-        /**
-         * The interval to execute the {@link dk.cloudcreate.essentials.components.foundation.messaging.queue.micrometer.DurableQueuesMonitor}s
-         * @return monitoring interval
-         */
-        public Duration getInterval() {
-            return this.interval;
-        }
-
-        public boolean isMonitoringDurableQueueSizes() {
-            return this.enabled && this.micrometerMonitorEnabled;
-        }
-
     }
 
 }
