+++ components/spring-boot-starter-postgresql/src/main/java/dk/cloudcreate/essentials/components/boot/autoconfigure/postgresql/EssentialsComponentsConfiguration.java
@@ -17,54 +17,31 @@
 package dk.cloudcreate.essentials.components.boot.autoconfigure.postgresql;
 
 
-import com.fasterxml.jackson.annotation.JsonAutoDetect;
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.MapperFeature;
+import com.fasterxml.jackson.annotation.*;
 import com.fasterxml.jackson.databind.Module;
-import com.fasterxml.jackson.databind.SerializationFeature;
+import com.fasterxml.jackson.databind.*;
 import com.fasterxml.jackson.databind.json.JsonMapper;
 import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import dk.cloudcreate.essentials.components.boot.autoconfigure.postgresql.EssentialsComponentsProperties.DurableQueuesProperties;
-import dk.cloudcreate.essentials.components.boot.autoconfigure.postgresql.EssentialsComponentsProperties.FencedLockManagerProperties;
-import dk.cloudcreate.essentials.components.distributed.fencedlock.postgresql.PostgresqlFencedLockManager;
-import dk.cloudcreate.essentials.components.distributed.fencedlock.postgresql.PostgresqlFencedLockStorage;
-import dk.cloudcreate.essentials.components.foundation.fencedlock.FencedLockEvents;
-import dk.cloudcreate.essentials.components.foundation.fencedlock.FencedLockManager;
-import dk.cloudcreate.essentials.components.foundation.json.JSONSerializer;
-import dk.cloudcreate.essentials.components.foundation.json.JacksonJSONSerializer;
-import dk.cloudcreate.essentials.components.foundation.lifecycle.DefaultLifecycleManager;
-import dk.cloudcreate.essentials.components.foundation.lifecycle.LifecycleManager;
+import dk.cloudcreate.essentials.components.boot.autoconfigure.postgresql.EssentialsComponentsProperties.*;
+import dk.cloudcreate.essentials.components.distributed.fencedlock.postgresql.*;
+import dk.cloudcreate.essentials.components.foundation.fencedlock.*;
+import dk.cloudcreate.essentials.components.foundation.json.*;
+import dk.cloudcreate.essentials.components.foundation.lifecycle.*;
 import dk.cloudcreate.essentials.components.foundation.messaging.RedeliveryPolicy;
-import dk.cloudcreate.essentials.components.foundation.messaging.eip.store_and_forward.Inboxes;
-import dk.cloudcreate.essentials.components.foundation.messaging.eip.store_and_forward.Outboxes;
-import dk.cloudcreate.essentials.components.foundation.messaging.queue.DurableQueues;
-import dk.cloudcreate.essentials.components.foundation.messaging.queue.DurableQueuesInterceptor;
-import dk.cloudcreate.essentials.components.foundation.messaging.queue.QueueName;
-import dk.cloudcreate.essentials.components.foundation.messaging.queue.QueuePollingOptimizer;
+import dk.cloudcreate.essentials.components.foundation.messaging.eip.store_and_forward.*;
+import dk.cloudcreate.essentials.components.foundation.messaging.queue.*;
 import dk.cloudcreate.essentials.components.foundation.messaging.queue.micrometer.*;
-import dk.cloudcreate.essentials.components.foundation.postgresql.MultiTableChangeListener;
-import dk.cloudcreate.essentials.components.foundation.postgresql.PostgresqlUtil;
-import dk.cloudcreate.essentials.components.foundation.postgresql.SqlExecutionTimeLogger;
-import dk.cloudcreate.essentials.components.foundation.postgresql.TableChangeNotification;
-import dk.cloudcreate.essentials.components.foundation.reactive.command.DurableLocalCommandBus;
-import dk.cloudcreate.essentials.components.foundation.reactive.command.UnitOfWorkControllingCommandBusInterceptor;
-import dk.cloudcreate.essentials.components.foundation.transaction.UnitOfWork;
-import dk.cloudcreate.essentials.components.foundation.transaction.UnitOfWorkFactory;
-import dk.cloudcreate.essentials.components.foundation.transaction.jdbi.HandleAwareUnitOfWork;
-import dk.cloudcreate.essentials.components.foundation.transaction.jdbi.HandleAwareUnitOfWorkFactory;
+import dk.cloudcreate.essentials.components.foundation.postgresql.*;
+import dk.cloudcreate.essentials.components.foundation.reactive.command.*;
+import dk.cloudcreate.essentials.components.foundation.transaction.*;
+import dk.cloudcreate.essentials.components.foundation.transaction.jdbi.*;
 import dk.cloudcreate.essentials.components.foundation.transaction.spring.jdbi.SpringTransactionAwareJdbiUnitOfWorkFactory;
 import dk.cloudcreate.essentials.components.queue.postgresql.PostgresqlDurableQueues;
 import dk.cloudcreate.essentials.jackson.immutable.EssentialsImmutableJacksonModule;
 import dk.cloudcreate.essentials.jackson.types.EssentialTypesJacksonModule;
-import dk.cloudcreate.essentials.reactive.EventBus;
-import dk.cloudcreate.essentials.reactive.EventHandler;
-import dk.cloudcreate.essentials.reactive.LocalEventBus;
-import dk.cloudcreate.essentials.reactive.OnErrorHandler;
-import dk.cloudcreate.essentials.reactive.command.CommandBus;
-import dk.cloudcreate.essentials.reactive.command.CommandHandler;
-import dk.cloudcreate.essentials.reactive.command.SendAndDontWaitErrorHandler;
+import dk.cloudcreate.essentials.reactive.*;
+import dk.cloudcreate.essentials.reactive.command.*;
 import dk.cloudcreate.essentials.reactive.command.interceptor.CommandBusInterceptor;
 import dk.cloudcreate.essentials.reactive.spring.ReactiveHandlersBeanPostProcessor;
 import io.micrometer.core.instrument.MeterRegistry;
@@ -73,24 +50,19 @@ import io.micrometer.tracing.Tracer;
 import io.micrometer.tracing.propagation.Propagator;
 import org.jdbi.v3.core.Jdbi;
 import org.jdbi.v3.postgres.PostgresPlugin;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.slf4j.*;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingClass;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.boot.autoconfigure.condition.*;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.event.ContextRefreshedEvent;
 import org.springframework.context.event.EventListener;
+import org.springframework.context.event.*;
 import org.springframework.jdbc.datasource.TransactionAwareDataSourceProxy;
 import org.springframework.transaction.PlatformTransactionManager;
 
 import javax.sql.DataSource;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 
 import static dk.cloudcreate.essentials.shared.FailFast.requireNonNull;
 
@@ -161,29 +133,7 @@ public class EssentialsComponentsConfiguration {
     @ConditionalOnProperty(prefix = "management.tracing", name = "enabled", havingValue = "true")
     public DurableQueuesMicrometerInterceptor durableQueuesMicrometerInterceptor(Optional<MeterRegistry> meterRegistry,
                                                                                  EssentialsComponentsProperties properties) {
-        var enabled = !properties.getDurableQueuesMonitor().isMonitoringDurableQueueSizes();
-        return new DurableQueuesMicrometerInterceptor(meterRegistry.get(), properties.getTracingProperties().getModuleTag(), enabled);
-    }
-
-    @Bean
-    @ConditionalOnMissingBean
-    public DurableQueuesMonitorManager durableQueuesMonitorManager(EssentialsComponentsProperties properties,
-                                                                   DurableQueues durableQueues,
-                                                                   List<DurableQueuesMonitor> monitors
-    ) {
-        var enabled = properties.getDurableQueuesMonitor().isEnabled();
-        var interval = properties.getDurableQueuesMonitor().getInterval();
-        return new DurableQueuesMonitorManager(enabled, interval, durableQueues, monitors);
-    }
-
-    @Bean
-    @ConditionalOnProperty(prefix = "management.tracing", name = "enabled", havingValue = "true")
-    public DurableQueuesMicrometerMonitor durableQueuesMicrometerMonitor(Optional<MeterRegistry> meterRegistry,
-                                                                         DurableQueues durableQueues,
-                                                                         EssentialsComponentsProperties properties) {
-        var enabled = properties.getDurableQueuesMonitor().isMicrometerMonitorEnabled();
-        String moduleTag = properties.getTracingProperties().getModuleTag();
-        return new DurableQueuesMicrometerMonitor(meterRegistry.get(), durableQueues, moduleTag, enabled);
+        return new DurableQueuesMicrometerInterceptor(meterRegistry.get(), properties.getTracingProperties().getModuleTag());
     }
 
     /**
